// src/app/domains/settings/constants/settings-sections.enum.ts
export enum SettingsSections {
    Profile = 'settings.profile',
    GovernmentAuthorities = 'settings.governmentAuthorities',
    VisitReports = 'settings.visitReports',
    Favorites = 'settings.favorites',
    Prescriptions = 'settings.prescriptions',
    InvitationCode = 'settings.invitationCode',
    Security = 'settings.security',
    Help = 'settings.help',
    Faq = 'settings.faq',
    ImportantNumbers = 'settings.importantNumbers.title',
    ContactUs = 'settings.contactUs',
} 